Update new_experiment_ui.py
Fixed Unequal group sizes. Previously, the warning would override the third warning label that would be raised when an experiment name was already in use. After believing I fixed the issue, the warning would simply not pop up. Now the issue is resolved and pops up for unequal group sizes.

Signed-off-by: rc10283 <99059695+rc10283@users.noreply.github.com>