Fix Remove RFID
Remove RFID just removes it from the GUI. The database still has it existing. When a new selection is placed after a RFID is removed, it proceeds to give the animal id to the number following the deleted rfid. So if an experiment has 10 animals, one is deleted, you are actually only able to have 9 animals in the experiment.

Signed-off-by: rc10283 <99059695+rc10283@users.noreply.github.com>